kms: fix problem with multiple displays
Previously, by trying to re-use the fb creation in _resize(), we'd end
up with a call back into _set_mode_major() with a bogus (all zero's)
mode.  Which results in trying to set mode 0x0@0 which results in
bad things.

The clocks react badly when trying to set them to 0Hz, so we need a bit
of protection in the kernel for this.  But we should get rid of the
bogus modeset in DDX as well (which this patch does).

Signed-off-by: Rob Clark <robdclark@gmail.com>
[tomasz.figa: Applied to xf86-video-openfimg]
Signed-off-by: Tomasz Figa <tomasz.figa@gmail.com>